# tiny formats changes to pty.py