the `__all__` list, did not had spaces betwen arguments, and one `os.close` had a space between the function name and  parenthesis and the  since every linter known to man think this 2 things are bad, i get warning when working with python code... and since i was doing this changes, i decide to f-string 2 extra lines... but i tried to keep changes to a minimum, so this PR is consider a typo/format/noop PR.