@aleivag 2019-02-27
hi @eamanu , i understand, but how about the `fstring` part, that can be consider a performance improvement (fstrings are faster than string `+` ) ?