@serhiy-storchaka 2019-02-28
Could you provide benchmarks that show the performance gain from this change?