@aleivag 2019-03-01
hi @serhiy-storchaka , sure i can, so, this is not the most comprehensive, or more scientific approach ever, but...

comparing master and my branch, on `pty._open_terminal`, using the worst case scenario (no file can be open, so we try all combinations).

# master

```
Python 3.8.0a2+ (heads/master:c606a9cbd4, Feb 28 2019, 22:33:12) 
[GCC 8.2.1 20181127] on linux
Type "help", "copyright", "credits" or "license" for more information.
>>> import timeit
>>> min(timeit.timeit("with suppress(Exception):\n    pty._open_terminal()", "from contextlib import suppress; import pty", number=1000) for _ in range(10))
1.5326147790001414
>>> min(timeit.timeit("with suppress(Exception):\n    pty._open_terminal()", "from contextlib import suppress; import pty", number=1000) for _ in range(10))
1.5294082229993364
```

# format-pty

```
Python 3.8.0a2+ (heads/format-pty-dirty:c606a9cbd4, Feb 26 2019, 22:47:55) 
[GCC 8.2.1 20181127] on linux
Type "help", "copyright", "credits" or "license" for more information.
>>> import timeit
>>> min(timeit.timeit("with suppress(Exception):\n    pty._open_terminal()", "from contextlib import suppress; import pty", number=1000) for _ in range(10))
1.5037514189971262
>>> min(timeit.timeit("with suppress(Exception):\n    pty._open_terminal()", "from contextlib import suppress; import pty", number=1000) for _ in range(10))
1.4546413489988481
```

i consistently get better times using f string that string concatenation.


to be honest i thought that this was more of known fact, in general (and specially true for simple types), fstring will be faster that any type of string formatting, and string concatenation.

now i want to give my disclosure, about this, i dont think this is a huge win, i just submitted this PR because my linter keeps complaining about the file formatting. and mention the "fstring performance win" is my backdor into getting this approved :P and mainly joking because of @eamanu  comment "
We don't pay attention of pep8 fixes unless there's an actual bug here". and all this in good conscience off course, and with the best intention.