@serhiy-storchaka 2019-03-03
`_open_terminal()` is a private function, which should never be called by the user directly. You should test public functions which use it.

In any case, I do not see significant difference (as expected).

Unpatched:
```
$ ./python -m perf timeit -s 'from contextlib import suppress; import pty' 'with suppress(Exception): pty._open_terminal()'
Mean +- std dev: 728 us +- 54 us
$ ./python -m perf timeit -s 'from contextlib import suppress; import pty' 'with suppress(Exception): pty._open_terminal()'
Mean +- std dev: 700 us +- 45 us
$ ./python -m perf timeit -s 'from contextlib import suppress; import pty' 'with suppress(Exception): pty._open_terminal()'
Mean +- std dev: 683 us +- 24 us
```
Patched:
```
$ ./python -m perf timeit -s 'from contextlib import suppress; import pty' 'with suppress(Exception): pty._open_terminal()'
Mean +- std dev: 708 us +- 43 us
$ ./python -m perf timeit -s 'from contextlib import suppress; import pty' 'with suppress(Exception): pty._open_terminal()'
Mean +- std dev: 707 us +- 35 us
$ ./python -m perf timeit -s 'from contextlib import suppress; import pty' 'with suppress(Exception): pty._open_terminal()'
Mean +- std dev: 719 us +- 30 us
```
If there is any difference, it is much smaller than the standard deviation. Sorry, I am closing this PR.